{{cookiecutter.project_name}}/requirements.txt
{%- if cookiecutter.include_pipeline == "y" -%}
mcot.pipe
{%- endif -%}